Just Some changes

+++ src/com/nativedevelopment/smartgrid/IPackage.java
@@ -6,7 +6,7 @@ import java.util.UUID;
 public interface IPackage extends Serializable{
 	public UUID GetRoutIdentifier();
 	public UUID GetCorrelationIdentifier();
-	public int GetFlag();
+	public long GetFlag();
 	public long GetTimeStamp();
 	// TODO maybe change return Serializable to bytes[]
 	public Serializable GetContent();
